Random dogs stored in App component
randogs data and functionality now stored in the App component.
randogs array and currentDog now stored in state. currentDog (url) is
passed down to the Randog component as a prop. Randog then displays an
image with this.props.currentDog.

On successful image upload, App currentDog state is updated to display
the most recently uploaded image. Randog is re-rendered to show the
image user just uploaded as confirmation of succesful upload.