Created elements for next and prior image
div elements are rectangles that are transparent and exist over the
left and right sides of the image. On hover the image background-color
will turn to blue and 25% opacity so user knows it is clickable.

nextImage and priorImage are functions that increment/decrement the
imageIndex state on the Wagger component unless it reaches the min/max

User is able to look through the images and select yes or no if it's
a match