Connect Yes and No buttons to API
Wagger component is able to update the API depending on
user input. Is it a match? Yes => create a like
and go to the next wagger (increment wag)
if No => go to the next waggger (increment wag)

Need to be able to recognize when theres a match,
and redirect to a new Match component